+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to Calendar Versioning (CalVer).
 
 ### Fixed
 - Fixed QR scanner not showing camera on first attempt after giving permission [PR #654](https://github.com/marmot-protocol/whitenoise/pull/654)
+- Fix show author name in last message of pending invites in chat list [PR #668](https://github.com/marmot-protocol/whitenoise/pull/668)
 
 ### Security